Add PHY synchronization block
Add block to handle the DVB-S2 PHY synchronization including:

- Frame timing detector and locking logic.
- Data-aided coarse and fine frequency offset estimation.
- Carrier phase tracking.
- PLSC decoding logic (interleaved first-order Reed-Muller code decoding).
- Coherent and non-coherent pi/2 BPSK demapping.
- Frequency correction logic based on an external rotator block.
- PL descrambler.
- Tagging of XFECFRAME start.

[Flowgraphs]

Add example flowgraphs for reception based on an IQ file, and for
loopback Tx/Rx simulation with AWGN. Add also a full Rx application.

[Frequency Recovery]

The data-aided frequency offset estimation is based on the full
PLHEADER (all 90 symbols). It is done after frame locking and proper
decoding of the PLSC. The latter ensure the PLSC symbols are known and
can be used as pilots for estimation (in addition to the SOF).

In the data-aided frequency offset estimation:

- Follow the approach of Ref 1 (indicated on the .h file) where the
  auto-correlation used for frequency offset estimation relies on
  accumulation among several consecutive frames.

- Support unlocked coarse estimation and correction. When unlocked, run
  the coarse frequency offset estimation algorithm based on the SOF
  symbols only. When locked, run over the full PLHEADER. Use the same
  strategy for estimating the PLHEADER phase for de-rotation. That is,
  estimate it with SOF symbols when unlocked and with the full PLHEADER
  symbols when locked.  Also, when the residual frequency offset isn't
  yet sufficiently low, de-rotate the PLHEADER symbols using the most
  recent coarse estimate prior to attempting PLSC decoding. This can
  improve the PLSC decoding performance.

Include message port to control an external rotator. Assume the PLSYNC
block receives the symbol-spaced stream out of a symbol
synchronizer/matched filter (MF) and that before the MF lies a rotator
for frequency correction. The new message port can control the frequency
correction applied by this assumed upstream rotator.

When controlling the external rotator, calibrate the delay between the
rotator and the plsync block. The plsync block can verify whether the
frequency update takes effect at the start of a PLHEADER by checking the
tag that the rotator places on the sample where the update starts. Use
this to calibrate the delay between the two blocks and ultimately for
the plsync block to schedule updates at the correct offset.

[Frame Timing Recovery]

After decoding the PLSC, compute the frame length and use that on the frame
synchronizer in order to check whether the frame timing is locked.

In frame time locking logic:

- Compute the correlation at the right time when locked. Within the 90
  symbols before that, only push symbols (differentials) into buffers,
  but do not compute correlation.

- Handle the unlocking logic. Use distinct threshold for frame timing
 unlocking. A distinct threshold specifically for unlocking is helpful
 in order to avoid unnecessary loss of frame timing lock.

[Carrier Phase Tracking]

Add naive data-aided phase tracking

Adjust phase of output symbols by a "static" phase that comes from the last
data-aided phase estimate.

[PLSC Decoding]

To decode PLSC, estimate PLHEADER phase using SOF. The de-rotation that
precedes the PLSC decoding is meant to improve the chances of decoding
the PLSC correctly. At this point, however, we can't estimate the phase
of the PLHEADER using the full PLHEADER symbols, because we don't know
what they are yet (the PLSC has not been decoded yet). Also, it does not
work to rely on the modcod info of the previous frame in case VCM is
being used. Hence, instead, estimate the PLHEADER's starting phase based
on SOF symbols only.

[PL Descrambler]

Regarding the PL descrambler, make sure that the pilots processed by the
fine frequency offset step must be descrambled. Output descrambled data
symbols only.

[pi/2 BPSK demapping]

Regarding pi/2 BPSK demapping, use coherent demapping after coarse freq
correction. Before coarse frequency offset correction is inferred,
i.e. before the residual frequency offset becomes within the fine
correction range, coherent pi/2 BPSK detection hardly works. In this
stage, it may be better to rely on differential (i.e., non-coherent)
detection of the PLHEADER.

[Tagging]

Tag every XFECFRAME start. The block following the PL sync block will
need to know the start of the XFECFRAME and the MODCOD used along it in
order to de-map the constellation symbols. The next block will also need
the MODCOD and the short frame indicator in order to decode the
interleaved FECFRAME. So tag this information in the PL Sync output.