Add dedicated and optimized module for pi/2 BPSK
The logic for pi/2 BPSK demapping was previously within the PLSC
decoder. For the sake of organization, separate it on a dedicated
source/header and reuse it on the PLSC module.

On the new dedicated module, add also a pi/2 BPSK mapping
function. Reuse it on the frequency synchronizer, which was previously
mapping PLHEADER symbols from scratch.

Optimize both mapping and demapping implementations.

Add also some unit tests covering both mapping and de-mapping functions.