Add double-ended queue and wrap it on delay line
The delay line was not sufficient. On the frame sync block, one of the
use cases is to store the symbols on a buffer such that, when the
PLHEADER start is detected, the PLHEADER symbols are readily
available. In this case, when reading the buffer from tail to head, the
tail should hold the oldest symbol (first PLHEADER symbol) and the head
should hold the most recent symbol (last PLHEADER symbol). This is the
reverse order in which the delay line stores samples.

Hence, add a more general double-ended contiguous queue implementation
that allows for storage in both orders (most recent to oldest and the
reverse). Change the delay line template class such that it inherits
from this new double-ended queue.