Fix last SOF symbol hard-coded on diff demapper
The previous implementation of the differential (non-coherent) pi/2 BPSK
demapper had the last SOF symbol hard-coded. However, because the whole
point of differentially decoding is that the symbols can be rotated, it
is better to use the actually received value of the last SOF
symbol. This patch uses this approach and adjusts various
comments/docstrings to clarify the expected BPSK symbols on the input to
differential detection functions (including the last SOF symbol).