Update the interface to decoded PL signaling info
The previous implementation held the PL signaling information directly
as public attributes of the plsc_decoder class. This patch changes the
interface such that there is a dedicated pls_info_t structure and the
plsc_decoder class holds a single such structure as a private
member. Now, to access the PL signaling information, the caller has to
copy it from the plsc_decoder class.

The goal with this patch is twofold. The first is to make sure the
caller/parent code does not tamper with the PL signaling information
accidentally. The second is to facilitate the tracking of multiple PL
signaling information corresponding to distinct frames. One can call
get_info() for each frame and copy the whole information structure.