Change the descrambler to descramble in one go
The previous implementation was meant to be used step-by step. That is,
the descramblign function would be called individually for each payload
symbol. This patch modifies the function to process the entire payload
in one go. This is going to be necessary once the code is refactored to
process the entire payload in one go.