Refactor the PL Sync processing strategy
Wait two SOFs before processing a PLFRAME payload, instead of processing
the symbols in real-time as they are fed on the input.

This approach is useful mostly because:

- It allows for an extra validation before processing the frame, that
  the frame length matches the interval between consecutive SOFs.

- It ensures the entire payload is ready to be processed in one go. With
  that, volk/SIMD commands can be used over the payload more easily.

Aside from that, the code looks a lot more organized and readable this
way.

This patch also updates the QA tests accordingly.