bbdeheader: add BBHEADER DFL/SYNCD validation
The previous implementation could deadlock on the main "while(h->dfl)"
loop in case the DFL or SYNCD fields of the BBHEADER were invalid. This
is a real possibility because the CRC-8 is not 100% guaranteed to detect
all errors. An error pattern that is a multiple of the generator
polynomial can result in undetectable errors and lead to a deadlock.

This patch adds some extra verifications on the DFL and SYNC values to
avoid deadlocking. It also adds QA test cases covering the referred
DFL/SYNCD error conditions.