Update workaround for Py tests before installation
The previous implementation included the "build/python/bindings"
directory within the PYTHONPATH and imported the cpython library named
"dvbs2rx_python" directly from this directory whenever the importing of
"dvbs2rx" failed in the pre-installation test environment. That was good
enough while the Python package didn't include any pure Python code. On
the other hand, now that there is pure python code in the package (i.e.,
params.py and defs.py), another workaround is required. For example, the
previous solution was not working for qa_params.py, which is a test that
imports the pure python code on params.py.

As a new workaround, include the entire build directory within the
PYTHONPATH instead of including the "build/python/bindings" directory.
Then, since the "build/python" directory has the "__init__.py" file,
import the "python/" directory from the "build/" directory as a package.
More specifically, call "import python as dvbs2rx".

To facilitate this workaround, include the build directory on the
PYTHONPATH that is set in the shell script generated for each test. With
that, it's no longer necessary to set PYTHONPATH globally on the test
environment, e.g., on the Github workflow environment.