Distinguish the open/closed loop freq sync modes
The previous frequency correction step applied by the frequency
synchronizer's `derotate_plheader()` method was not suitable for
operation in closed-loop, with an external derotator. The internal and
external corrections could easily conflict with each other. This patch
enhances the documentation regarding the open-loop/closed-loop use cases
and adds an open-loop option on the referred method.