PL Sync: fix verification of tag offset error
It's the instantaneous error that matters, not the cumulative error (the
actual delay). That's an important distinction because the cumulative
delay can increase over time, due to how the symbol sync block operates.
Hence, the cumulative delay easily exceeds the tolerance, while the
instantaneous offset error should always be low if we are estimating the
delay correctly.