Review the API for pilot-mode fine freq estimation
The main goal of this match is to move all the interaction with the
"angle_pilot" buffer into a single place, the estimate_fine_pilot_mode
function. This is better to avoid errors with missing calls to set a
specific pilot phase or PLHEADER phase. Let the fine estimation
procedure handle all that.

In particular:

- Provide both the PLHEADER and descrambled payload to the fine
  frequency offset estimation function.
- Let the function compute the PLHEADER phase and pilot phases, and fill
  the "angle_pilot" buffer accordingly.
- Change the workaround for the anomaly on the first interval between
  the PLHEADER and the first pilot block (slightly larger than the
  inter-pilot intervals). Take the last 36 symbols of the PLHEADER only
  to preserver the 1476-symbol interval.