plsync: Fix rotator control scheduling using float
The target index where the external rotator should be adjusted comes
from the multiplication between sps (a small number) and the absolute
sample index (a large uint64_t). The previous implementation was quickly
losing precision significantly due to using sps as a single-precision
floating-point number. To solve the problem, cast sps to a double in the
computation.