Support PLFRAME filtering based on the PLS info
Add PLS filtering support in two parts: on the PLSC decoder, and on the
PL Sync processing.

On the PLSC decoder, use a priori knowledge about the set of possible
PLSC codewords in order to reduce the search for the minimum distance
between the input and reference codewords. This strategy should benefit
both the CPU usage (less codewords to test) and the decoding
performance, since the decoder won't return results that are not
possible according to a priori knowledge. At the same time, keep it
possible to disable the filters completely and let the decoder return
any result from the 128 possible PLSC codewords (ACM/VCM mode).

On the PL Sync block, the PLS filtering is used when deciding whether or
not to output a XFECFRAME. The block now outputs only the XFECFRAMEs
associated with PLS values that are enabled on the filters. This is
critical to preserve the boundaries between XFECFRAMEs on the downstream
blocks. For instance, if the LDPC decoder is configured to decode
QPSK1/4 with normal FECFRAME, then a single erroneous QPSK3/5 frame
would break the alignment considered by the LDPC decoder block. With the
PLS filters, however, we prevent this situation by rejecting PLFRAMEs
with unexpected PLS.