Review the default RRC filter delay on dvbs2-rx
The RRC filter length substantially affect the CPU utilization,
especially since the symbol synchronizer block (and the underlying
filtering) typically is the most CPU-hungry block. Hence, reduce the
default delay to a much lower value to make the filter shorter.

On the Tx side, it is important to keep the RRC delay quite high so that
the RRC sidelobes are low enough. In contrast, on the Rx side, it's ok
not to care that much about sidelobes and instead focus on the CPU
utilization.