Review the default coarse freq estimation period
The current implementation is a little sensitive to the coarse frequency
offset estimation. If the coarse estimate ever falls out of the
"coarse-corrected" range, the frequency correction loop will subtly
switch the correction value back from the fine estimate to the coarse
estimate, and most likely this will cause some data loss. To avoid this
situation, it seems preferable to make sure the coarse frequency offset
estimation averages out a higher number of PLFRAMEs by default.