Fix hard-coded history on symbol sync block
If we ever change the interpolator to a quadratic or cubic scheme, which
require a different sample history, make sure the interpolator can be
swapped easily while changing a single history variable. In other words,
make sure the history-dependent definitions actually use the current
history setting from a constant, instead of hard-coded definitions that
assume a history of 2.