bbdeheader: Do not process byte-misaligned SYNCDs
The processing loop implemented by the bbdheader block assumes the range
of (DFL - SYNCD) bits comprises an integer number of bytes. Otherwise,
if the input BBFRAME has a non-byte-aligned SYNCD value, the loop can
spin indefinitely. Prevent this scenario by discarding any BBFRAME
carrying a byte-misaligned SYNCD (non-multiple of 8).

Note it makes no sense to expect byte-misaligned SYNCDs. The BCH
dataword length (kbch) is always a multiple of 8, and the block supports
MPEG TS user packets only, which are also byte-aligned (188 bytes each).
Thus, the Tx can always include an integer number of bytes on the
DATAFIELD, so there is no way for a partial byte to span two DATAFIELDs.
Instead, a byte-misaligned SYNCD is more likely to be due to a corrupt
BBHEADER undetected by the CRC check.