symbol sync: Default to OOT interpolator on Rx app
The OOT implementation is faster and recommended for supported
oversampling ratios (even and integer numbers). The in-tree
implementation is more generic and recommended when the goal is to use a
fractional or odd oversampling ratio. Use the OOT implementation by
default and make sure the dvbs2-rx app instructs the user to pick the
in-tree implementation when appropriate.