plsync: Correct the fine freq. offset internally
The previous implementation relied exclusively on the external rotator
for the frequency correction (both coarse and fine). The rationale was
that eventually the correction would converge to a sufficiently accurate
value leaving a negligible residual frequency offset. In this
hypothesis, it would be unnecessary to waste resources correcting the
frequency offset again internally on the PL Sync block.

Nevertheless, the main limitation of this approach is that the
correction applied by the external rotator is not immediate. Instead, it
has a two-frame delay for the fine estimate. In other words, the
correction due to the fine frequency offset estimated on frame N only
takes effect at the start of frame N+2. Thus, it would only be possible
to rely exclusively on the external rotator if we could assume that the
frequency offset remains sufficiently stable in the course of two or
more frames. However, it turns out this assumption is often false.

To solve the problem, introduce a further frequency correction layer
internally to the PL Sync block. Implement it as a feed-forward
correction, which focuses on the current frame only and does not affect
the closed-loop mechanism involving the external rotator. See the
comments in the code.