dvbs2-rx: Default to SIS instead of auto SIS/MIS
The CCM/SIS configuration is by far the most common for digital TV and
other broadcasting applications. Since the PL Sync block performs better
when it knows that multistream is either off or on (instead of auto),
configure it to multistream=off by default. With that, it only takes the
--pilots=on/off setting to disable the PLSC decoder. In contrast, before
this patch, the user would need to set both the pilots and multistream
options to a non-auto value in order to disable the PLSC decoder.