Remove level-based vector dumps used for design
They were useful when implementing the frame/freq sync algorithms, but
they are mostly a tool for developers/designers, not users. Remove them
for now to avoid the performance cost imposed by the conditionals around
each dump. Besides, if we ever need to troubleshoot an algorithm in the
field, we can fetch an IQ recording and analyze offline without
resorting to such vector dumps.