Add BCH benchmarking with a comparison to Aff3ct
Add an option to run the benchmarking both with the current gr-dvbs2rx
BCH implementation (and the BCH encoder from GR in-tree) as well as the
Aff3ct implementation. At the moment, the Aff3ct implementation produces
different results due to the order of systematic/parity bits and other
possible issues likely related to bit order. Hence, as of this patch,
the GR encoder is used with the gr-dvbs2rx decoder, while the Aff3ct
encoder is used with the Aff3ct decoder only. That is, it is not
possible to run the GR encoder and decode with the Aff3ct decoder.