dvbs2-rx: Fix GUI incompatibility with GR 3.10
The GR 3.10.1 package from the ppa replaces pyqwidget() with qwidget(),
which breaks the GUI instantiation on dvbs2-rx. Fix this by checking
whether pyqwidget() is available on the QT gui object and calling
qwidget() otherwise. For better code reuse, wrap this logic into a
function and reuse it for all QT GUI blocks.