dvbs2-tx: Add GUI mode with USRP controls
Implement GUI mode on the Tx application. By default, add labels with
the main DVB-S2 parameters on the GUI. If the sink is a USRP device, add
the gain and frequency controls (GUI range widgets). Support also
optional plots for observing the output IQ in the time and frequency
domain.