rx: Fix option for PLSC decoding in VCM/ACM mode
The --pl-acm-vcm option is currently mainly intended as a tool to
identify the MODCOD/frame/pilot configuration of a target unknown
signal. Hence, it should operate with all-pass PLS filters on the PL
Sync block. The previous implementation was providing the ACM/VCM flag
to the PL Sync block correctly, which was enough to include the
possibility of dummy PLFRAMEs. On the other hand, it was restricting the
PLS filters to the chosen MODCOD only, so the PL Sync block was not
considering all possible PLS values during PLSC decoding. Fix this by
providing the all-pass PLS filters instead.