plsync: Map requested rotator phase inc changes
Put each phase increment adjustment request (sent via message port) on a
map and match each incoming rotator phase inc adjustment tag with the
requested update. This is a better approach to avoid problems with
duplicate tags or processing tags associated with a distinct PLFRAME. It
also holds more information to allow for more precise warnings.

Add an extra piece of information on the rotator phase inc adjustment
tag to help with the matching between a tag and a map entry. More
specifically, add the originally requested phase increment update
offset. The rotator block includes the information within the tag value
and the PL Sync block reads it to identify the corresponding request on
its update map.

As a minor update, change the semantics of the rot_freq_t structure.
Rename it to rot_state_t (given it refers to the rotator state) and save
the rotator's rotating frequency on it instead of its negative (the
corrected frequency offset).