Fix rotator_cc importing on rotator QA
The tests are designed to be executable before dvbs2rx is installed in
the system. To do so, we add a fallback import statement to import
dvbs2rx directly from the python folder within the build directory when
the global gnuradio.dvbs2rx is not found. This statement was missing on
the rotator QA.