Build with native optimizations on by default
It seems that the need/usefulness of native optimizations comes more
frequently than the scenarios where we don't want them. We don't want
them enabled when building binary packages or a Docker image to be
distributed. Hence, leave the option on by default and disable it when
necessary. Also, add an option (a Dockerfile argument) for disabling it
on the Docker image build.