Avoid symbol_sync test errors due to scipy upfirdn
The upfirdn function used for upsampling on the symbol_sync's open-loop
test may disturb the results. In old scipy.signal versions, upfirdn (in
the way it was called) results in a sequence with "sps * len(x_syms)"
samples, with "sps - 1" zeroes in between each symbol. Namely, the same
as "upsample(x, sps)" on MATLAB. However, scipy version 1.8 (and
possibly many other close versions) seems to cut the trailing zeroes on
the upfirdn result, so the last symbol is not followed by "sps - 1"
zeroes. With that, the number of samples processed by the symbol sync
loop would change and the test would fail as reported in Issue #11.

To avoid uncertainty on this test case, apply the upsampling manually
with a simple for loop. There is no need to use upfirdn for that.