Move Rx hier block to the examples directory
The Rx hier block is currently used by the examples only, and not by the
main dvbs2-tx/rx apps. Instead, the main apps connect the subblocks one
by one to allow for various customizations. As a result, the hier block
is currently optional just like the example flowgraphs, which means the
cmake build doesn't need to generate the hier block automatically.

By moving the hier block to the examples/ directory, any flowgraph using
it will trigger its generation when compiled. With that, we avoid
installation problems like the one reported in Issue #2.

Fixes #2.