tx: Default IQ full scale to 1.0 on fd/file sink
Previously, the full scale was only limited when the Tx sink was the
USRP. However, the fd/file sink can be used to feed a Tx radio too (via
pipe), so the previous full scale could cause a problem. To avoid that,
default the full scale to 1.0 in all cases and add option to disable the
full-scale limit.