tx: Fix RRC tap scaling for a target IQ full scale
The previous implementation scaled the RRC taps so that the output IQ
magnitude would range within +-fullscale. However, SDR Tx interfaces
feed I and Q samples into separate DACs, so the full scale should apply
to the individual I and Q amplitudes, not the IQ magnitude. This patch
adjust the scaling by a factor of sqrt(2) so that the I and Q samples
are instead within +-fullscale.