plsync: Fix expected PLSC codewords optimized out
As of gcc 13.1.1, as tested on Fedora 38, the list of expected PLSC
codewords computed by the PL Sync block was wrong due to a compiler
optimization of the (1ULL << pls) mask computed in the loop. This
problem affected release builds only. This commit replaces the
implementation with one that does not get optimized out.