Review leak bytes on LUT-based GF2 poly remainder
Update the number of leak bytes from the degree of the divisor
polynomial to the maximum number of bytes that fit in T. We can choose
any number of leak bytes in advance, as long as it is greater than or
equal to the divisor's degree. Hence, use the maximum number of bytes
that can be processed per iteration instead of the minimum.