Change u8-array-based syndrome to pointer argument
It is better to pass the codeword by a u8 array pointer (u8_ptr_t) than
a u8 vector because the codeword may not be in a vector, but rather a u8
buffer. For instance, with GNU Radio scheduler buffers, the codeword
would be on a u8 buffer, so copying to a u8 vector would be unnecessary.