Support BCH decoding from array of bytes
When n and k are multiples of 8, the message and codewords can be easily
represented by u8 arrays. This patch adds the missing function to decode
a codeword expressed by an array of bytes, with the logic to correct
errors in the array and return the message as another array of bytes.