Use the new BCH codec on the BCH decoder block
Replace the old implementation with the decoder from the new BCH codec
implementation. Also, move the old sources to the benchmarking directory
so that the old implementation remains available on benchmarking.

Since the new implementation works with bit-packed bytes instead of
unpacked bytes, this patch requires compatibility updates to the LDPC
decoder, BB descrambler, and BB deheader blocks, which will follow.