Review importing from build dir on Python QA tests
Remove the previous appending of the bindings dir on Python QA via
"sys.path.append". The goal in doing so was to be able to find the
dvbs2rx_python.cpython shared library on the build dir. However, this
did not produce any effect because the sh script created for the test
runs the test directly from its source directory, not from the build
directory. Hence, the bindings dir that ended up being added was the
source bindings directory, which does not contain the
dvbs2rx_python.cpython library.

Despite the above, the tests still worked from the build dir due to the
GR_TEST_ENVIRONS setting on CMakeLists.txt, which adds the build dir to
PYTHONPATH on tests. With that, any import of python.dvbs2rx finds the
package on the build dir. Then, when __init__.py is loaded (from the
build dir), calling "__path__.append(os.path.join(dirname, "bindings"))"
finally adds the build/python/dvbs2rx/bindings dir containing the
dvbs2rx_python.cpython shared library.

The same effect can be achieved with the new implementation in a cleaner
way. Now, __init__.py tries to load from ".bindings.dvbs2rx_python"
first, which works when the package is being loaded from the build dir
at, e.g., build/python/dvbs2rx. No manual addition of the bindings dir
to the sys path is needed in this case.