Skip correction if error-loc polynomial degree > t
This is a minor optimization. A codeword with uncorrectable errors not
necessarily yields an error-location polynomial with degree greater than
t, so this change does not affect all uncorrectable codewords.

Conversely, when the error-location polynomial has degree greater than
t, we know that the error-location numbers can be found, so some or all
errors will remain. Thus, we can skip the correction entirely to avoid
the expensive computation of error-location numbers.