WIP: Add debugging logs for custom bitset256
The BCH QA is still failing when using the new custom bitset256 type and
compiling in Release mode. In contrast, it does work when compiling in
Debug mode. This patch adds useful logs for troubleshooting the
discrepancy on the encode step when called from the
test_bch_dvbs2_encode_decode test case.