Add options to benchmark CPU usage on iq-rec-cli
Adds the --cpu and --plot-cpu command-line flags to enable the
measurement of CPU utilization per thread when running the rx or bench
iq-rec-cli commands. With the --cpu option, the utilization is printed
to the console in the end. With the --plot-cpu option, the tool also
generates a time-series plot with the CPU utilization per thread over
the duration of the IQ recording processing.