Decouple type in GF(2) poly conversion to GF(2^m)
The type used to store the polynomial over GF(2) is irrelevant when
converting it to a polynomial over GF(2^m). As long as we can identify
whether each bit is set, we can convert the polynomial over GF(2) to a
polynomial over GF(2^m) of arbitrary degree, for any m and type T used
to store the elements of the GF(2^m) field.

Previously, the conversion function (gf2m_poly constructor) assumed
gf2_poly<T> and galois_field<T>, meaning the same type T would be used
to store the polynomial over GF(2) and the elements of the GF(2^m)
field. That is not required at all, so decouple the two types.