Add GF2 poly remainder for distinct-type operands
Replace the former operator% of GF2 polynomials with a new non-member
remainder function that supports the computation for polynomials over
GF(2) using different types to store the coefficients. That is, support
a(x) % b(x) for a(x) represented by a gf2_poly<Ta> and b(x) represented
by a gf2_poly<Tb>, where the Ta and Tb types can be distinct.