Do not compute syndrome when r(x) % g(x) is zero
This is a performance optimization. A zero remainder obtained after
dividing the received codeword r(x) by the generator polynomial g(x)
implies the received codeword is error-free. In this case, there is no
need to evaluate the syndrome components, as they would all be zero
anyway. Instead, return an empty vector and consider the empty vector as
an indication of error-free codeword in the caller.

This optimization is particularly useful when the SNR is sufficiently
high for the adopted MODCOD such that errors in the output of the LDPC
decoder are rare.