Move XFECFRAME demapping to an independent block
Instead of having the XFECFRAME demapping logic within the LDPC decoder
block, move it to a standalone block. The challenge is handling the SNR
measurement. By having the demapping logic inside the LDPC decoder
block, it is easier to measure the post-decoder SNR. In contrast, with
the new approach, the LDPC decoder needs to send the decoded LLRs to the
XFECFRAME demapping block so that it can estimate the SNR.

The adopted solution is based on PDU ports. The LDPC decoder block sends
a PDU with the decoded LLRs back to the XFECFRAME demapper block. Hence,
this patch adds an output PDU port on the LDPC decoder block and a
corresponding input PDU port on the XFECFRAME demapper block.

Also, with this change, the LDPC decoder now takes int8_t on the input
instead of complex symbols. I.e., it receives the soft demapping
decisions out of the XFECFRAME demapper. Hence, the LDPC decoder block
suffix was renamed from _cb to _bb.

Moreover, since it is now the XFECFRAME demapper who estimates the SNR,
this patch removes the SNR fetching method from the LDPC decoder block
and moves it to the XFECFRAME demapper block. Correspondingly, it
updates the call on dvbs2-rx.

Process all LLRs