pkg: Fix dependency on deprecated distutils
The distutils package has been removed in Python 3.12. As a result,
starting on Fedora 39 (with py3.12), RPM packages would fail to build.
The same problem will eventually happen on other distros when they adopt
py3.12. To avoid future incompatibility, drop the current workaround for
defining GR_PYTHON_DIR using distutils and instead use the default
definition from GR's GrPython module. At the same time, set the paths
explicitly on the Deb and RPM packages so that they don't need to rely
on GrPython defining the paths correctly. In particular, on the Deb
package, make sure the path is set to dist-packages instead of
site-packages as otherwise would be set by GrPython.