fix: remove bogus sampling rate check
The plutosdr only accepts a value of datatype integer as sampling rate.
For this, there already is a conversion logic, to convert float data
type sampling rate values to integer data type. However, the conversion
logic is flawed, as it only triggers for values of float data type
**if the floating point number has decimals** (digits behind the dot).
Therefore, values of floating point data type like `10` or `25000`
are not converted to an integer data type, resulting in failure to
initialize the plutosdr.

This patch removes the bogus check, always converting to integer type.

This is a fix to https://github.com/igorauad/gr-dvbs2rx/issues/38 /